eslint fixed errors on admin.js and small refactoring of functions
