+++ public/js/shared/ajax.js
@@ -1,0 +1,17 @@
+function getDetails
+(
+    link,	
+    values,
+    requestType = "POST",
+    dataType = "json",
+    headers = {"X-CSRF-TOKEN": $('meta[name="csrf-token"]').attr("content")}
+) 
+{
+    return $.ajax({
+        headers,
+        url: link,
+        type: requestType,
+        dataType,
+        data: values,
+    });
+}
